refactor: move type_hash to attribute
