+++ packages/macros/src/attribute/mod.rs
@@ -1,3 +1,4 @@
 //! OpenZeppelin attribute macros.
 
+pub mod type_hash;
 pub mod with_components;
