+++ packages/macros/src/attribute/type_hash/definition.rs
