+++ packages/macros/src/attribute/type_hash/diagnostics.rs
