+++ packages/macros/src/attribute/type_hash/mod.rs
