+++ packages/macros/src/attribute/type_hash/parser.rs
