+++ packages/macros/src/attribute/type_hash/types.rs
