--- packages/macros/src/derive/mod.rs
@@ -1,3 +1,0 @@
-//! OpenZeppelin derive macros.
-
-pub mod type_hash;
