+++ packages/macros/src/lib.rs
@@ -1,9 +1,8 @@
 pub(crate) mod attribute;
 pub(crate) mod constants;
-pub(crate) mod derive;
 pub(crate) mod utils;
 
-pub(crate) use derive::type_hash;
+pub(crate) use attribute::type_hash;
 
 #[cfg(test)]
 mod tests;
